Пожелаем удачи фронтам! Топовая версия топового приложения Number One

+++ project/src/main/java/caselab/service/attribute/AttributeService.java
@@ -3,22 +3,31 @@ package caselab.service.attribute;
 import caselab.controller.attribute.payload.AttributeRequest;
 import caselab.controller.attribute.payload.AttributeResponse;
 import caselab.domain.entity.Attribute;
+import caselab.domain.entity.enums.GlobalPermissionName;
 import caselab.domain.repository.AttributeRepository;
-import java.util.List;
-import java.util.Locale;
-import java.util.NoSuchElementException;
+import caselab.exception.entity.not_found.AttributeNotFoundException;
+import caselab.service.util.PageUtil;
+import caselab.service.util.UserUtilService;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.MessageSource;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class AttributeService {
 
+    private final UserUtilService userUtilService;
     private final AttributeRepository attributeRepository;
-    private final MessageSource messageSource;
 
-    public AttributeResponse createAttribute(AttributeRequest attributeRequest) {
+    public AttributeResponse createAttribute(AttributeRequest attributeRequest, Authentication authentication) {
+        userUtilService.checkUserGlobalPermission(
+            userUtilService.findUserByAuthentication(authentication), GlobalPermissionName.ADMIN);
+
         Attribute attribute = new Attribute();
         attribute.setName(attributeRequest.name());
         attribute.setType(attributeRequest.type());
@@ -28,41 +37,45 @@ public class AttributeService {
 
     public AttributeResponse findAttributeById(Long id) {
         Attribute attribute = attributeRepository.findById(id)
-            .orElseThrow(() -> new NoSuchElementException(
-                getAttributeNotFoundMessage(id)
-            ));
+            .orElseThrow(() -> new AttributeNotFoundException(id));
         return new AttributeResponse(attribute.getId(), attribute.getName(), attribute.getType());
     }
 
-    public List<AttributeResponse> findAllAttributes() {
-        List<Attribute> attributes = attributeRepository.findAll();
-        return attributes.stream()
-            .map(attribute -> new AttributeResponse(attribute.getId(), attribute.getName(), attribute.getType()))
-            .toList();
+    public Page<AttributeResponse> findAllAttributes(
+        Integer pageNum,
+        Integer pageSize,
+        String sortStrategy,
+        Authentication authentication
+    ) {
+        userUtilService.checkUserGlobalPermission(
+            userUtilService.findUserByAuthentication(authentication), GlobalPermissionName.ADMIN);
+
+        PageRequest pageable = PageUtil.toPageable(pageNum, pageSize, Sort.by("name"), sortStrategy);
+        Page<Attribute> attributes = attributeRepository.findAll(pageable);
+        return attributes
+            .map(attribute -> new AttributeResponse(attribute.getId(), attribute.getName(), attribute.getType()));
     }
 
-    public AttributeResponse updateAttribute(Long id, AttributeRequest attributeRequest) {
+    public AttributeResponse updateAttribute(Long id, AttributeRequest request, Authentication authentication) {
+        userUtilService.checkUserGlobalPermission(
+            userUtilService.findUserByAuthentication(authentication), GlobalPermissionName.ADMIN);
+
         Attribute attribute = attributeRepository.findById(id)
-            .orElseThrow(() -> new NoSuchElementException(
-                getAttributeNotFoundMessage(id)
-            ));
-        attribute.setName(attributeRequest.name());
-        attribute.setType(attributeRequest.type());
+            .orElseThrow(() -> new AttributeNotFoundException(id));
+        attribute.setName(request.name());
+        attribute.setType(request.type());
         attribute = attributeRepository.save(attribute);
         return new AttributeResponse(attribute.getId(), attribute.getName(), attribute.getType());
     }
 
-    public void deleteAttribute(Long id) {
+    public void deleteAttribute(Long id, Authentication authentication) {
+        userUtilService.checkUserGlobalPermission(
+            userUtilService.findUserByAuthentication(authentication), GlobalPermissionName.ADMIN);
+
         if (attributeRepository.existsById(id)) {
             attributeRepository.deleteById(id);
         } else {
-            throw new NoSuchElementException(
-                getAttributeNotFoundMessage(id)
-            );
+            throw new AttributeNotFoundException(id);
         }
-    }
-
-    public String getAttributeNotFoundMessage(Long id) {
-        return messageSource.getMessage("attribute.not.found", new Object[] {id}, Locale.getDefault());
     }
 }
